requirements.txt
Pillow
drawnow
matplotlib
tqdm